Update issue 1540
Removing a unnecessary call that was clipping and projecting the wind energy
points from the biophysical run to the same AOI. This was an artifact left over
from the bio/val merger. Also, resolved a bug where opening the wind energy
points file was being done as read only and was throwing errors. Now open it in
writeable mode.

Also now pass in uri names including extensions to the clip_and_reproject_maps
function. This was causing a bug too. This function should really be two
functions.


Former-commit-id: 33218442bd2dacc71442e6dce3e726e0adf2d814