Update issue 1230
Docstring for make_risk_raster. It will now take in crit_lists and denoms. E and C combination will be done as subfunctions, and will be farmed out and returned as a single E/C raster for each. This raster can then be sent to array or something, and then use the risk_eq on it.


Former-commit-id: 0096be54021825092aeaa90de6f30ed2d3430a17