Update issue 1463
Following the suggestion of Luke Pinner (via my question on
gis.stackexchange.com), I'm adding a function to get the value under a
pixel at a given point defined by a shapefile.  This function appears to
work beautifully, but will need to be adapted to meet my needs for
pollination.


Former-commit-id: 2ee10bffed7221418eae58b401f7f9a18d95a111