Update issue 1487
The computational steps that adjust the ranks of a layer if it's specified by a uri, and the step that replaces a layer with a constant if its uri is not specified have been restored. They are both placed in the execute-core loop. I've added a string that encodes a type for each output, so now I have a way to know what output is a layer, and adjust the output's rank if necessary.


Former-commit-id: 179fe593ed76f3cf0421fe44de88e093708d069e