Update issue 1230
Fixed an issue in which the decayed raster was being unfairly pre-multiplied against the pre-rasterized H/S/H-S rasters. Instead of using it as a stencil, was actually being ,multiplied in. Need to only do r/dq*w where that H-S overlap or the H or the S exist, ut don't want them to have influence until the end when they get multiplied before the final risk raster is output.


Former-commit-id: 148a794633b1b9d0f1bb5956eef81bdee585b84b