Update issue 1230
Created the euclidean risk function. Since E and C are now both arrays, we have to preform the typical operations, but also make sure on the last add of e_array and c_array that we are only doing so if there is data in e_array (which was multiplied against the decayed raster, and thus would only contain data if there was decayed data in that area). We do this by masking e_array, then multiplying that against c_array, and when I know that their 0's will line up, can add them and then do the final square root.


Former-commit-id: a9ed3c022d2db817036b2c6217867d307130543e