Update issue 1230
Created the call to make_cum_risk_rasters. This will return a list of all of the sumulative habitat risk rasters so that it can be passed directly on to the make_ecosystem_raster function which which will take all habitat rasters and merge them into one final risk raster.


Former-commit-id: 9c5c33c48a2b3e4b5158a50e72540a36b5a7f44c