Update issue 1555
Modifying code so that I can test the computation of sheltered shoreline segments. There is a problem with the fetch algorithm: except for the exactly vertical or horizontal lines, it always return a bit less than max_fetch because of the discrete nature of the rasters. Blocking this issue on issue 1556 that is supposed to solve this problem.


Former-commit-id: e45605d9f72432bd0362e02b96fa6508dd808bc2