Update issue 1494
I came across a bug where the wind energy json data file was not being added to
the directory. James and I deduced that it should be moved outside of the
conditional for Operating System. James also figured out that the lib_path
needed to be concatenated to the location for where we wanted to store the file


Former-commit-id: b29867d4edc272dad61c194b440d03c3f64e9f17