Update issue 1556
Fixed a bad design decision in the computation of the fetch: I was using the ray's endpoints on the grid (int) to compute its length, instead of the actual coordinates (float) in both the slow and fast algorithms. Also, if the ray hits land before it reaches the maximum fetch distance, the algorithm returns a fetch that extends to the edge of the last water cell.


Former-commit-id: 019833bfd25b6d12741736c2b1a4c2202efdaf84