Update issue 1556
The test suite for the fetch pass. I couldn't get the computed values agree more than 13 decimal places after the coma, though. This has to do with the way the data is saved in the CSV file, because the data that is loaded and saved is identical.
Documented the portion of code used to generate the fetch distance/depth regression data for future use. Improved the comments in 'compute_fast_fetch'.


Former-commit-id: ce22a398822ffda34b765da9064dd1cd86f458f0