Update issue 1310
I think I've figured it out!

The issue with previous attempts at implementing tests in the UI is that
some functionality (notably the validation) happens in a thread outside
of the main thread of control.  When this happens, I need a way to pause
the test itself, but still leave the actual UI completely functional.

Enter the QTest class (PyQt4.QtTest.QTest), which provides the QTest.qWait
function.  this function allows the UI to proceed aside from the test itself.

the container testing class in iui_test.py contains an example of usage
for the qWait function.


Former-commit-id: 595cc2f0e1fe67f7c87af9b4e9a3eea38e0038de