Update issue 1310
Since the UI object heiarchy has so many subclasses, I thought it might
be useful to implement something like java's interfaces, where certain
functionsare required to be implemented.

With these couple of classes, it makes it very clear which functions
must be implemented at minimum.

I still need to think about the best way to run a battery of tests against
multiple UIs, of that is even necessary


Former-commit-id: b1d5039ec7406074479a56dcc13c1907c5af7a34