Update issue 1565
Added args['max_rating'] an an input to the non-core args dictionary. We want to get from the user the highest potential value that they want for the rating/data quality/ weights in their CSV. This will allow us to calculate the max possible risk to know what the percentage of total risk each individual risk pixel is.


Former-commit-id: 0134474f6bb977125cf5aafd6db7a1cd15f30d3e