Update issue 1565
Wrote a function to return the highest possible risk for any combination of h-s, given the risk equation and the maximum rating for a criteria within this model run. This will be passed along to the core in order to determine total potential risk for a habitat. This can be calculated in core by multiplying args['max_risk'] by the number of applicable stressors to that habitat.


Former-commit-id: 9ec2f28d01a4073708a2239ff1dfce71bb8c8c97