Update issue 1578
Adding a skeleton framework for setting up a standard modelUI instance and
setting the workspace value to something we want, like a predefined
test workspace.  The only assertion that happens right now is a check to
verify that there is no validation error once we set the workspace folder
to a new value.


Former-commit-id: 9b42aee79b1daa39d7e4552e730edc59fe3409f8