Update issue 1578
Making a small yet significant change to the execution of the operation
dialog (the 'running the model' modal window).  The change is to switch from
a function that shows the window but blocks until it returns (exec_()), to
a non-blocking function that still shows the window (show()).

This change causes my test to behave as expected, but I also expect that it
might cause problems elsewhere in an instance of the UI.


Former-commit-id: 8eaaf6fc84147d9aad523fa0dd06bde0eba28b6d