Update issue 1555
Fixed an issue in the computation of wave exposure where the depths were expected to be positive, but were passed to the equation as negative values. Uncommented all the inputs so that test_execute runs over all the inputs and produces all the intermediate and output files. Started to edit the code in anticipation of PyLint.


Former-commit-id: fd4c42ff3ad7fb6f381b19595ffa6a24a21840c8