Update issue 1230
No longer needed to do a lambda expression in order to get the unique habitats and stressors. This function already contains a dictionary in which rest keys that are exactly that. so instead, we can just pull the .keys() from the dictionary.

Should take a look at whether it's cheaper to do 2 lambda expressions later, or pre-create a list of both habitats and stressors.


Former-commit-id: 7124ec08d6bcd4e8a4981e9f0cc5c6428b10921d