Update issue 199
The raster_util's slope calcluation was incorrect, keeping the slope in terms of a fraction rather than percent.  So if the slope of the pixel was 10%, the function had reported 0.1, but many equations (like SDR) assume a 10% slope means the value plugged in will be 10.  Fixed the function and updated the regression test.


Former-commit-id: 4ddbb15cbc6f167980d891189871239e015610f6