update issue 1230
Needed to change the return of make_hab_risk_rasters from a straight list of the rasters to a dictionary mapping name to the open dataset. This will allow us to retrieve the names for use when making the percentile shapefile outputs.

Instead, are now lambda mapping the dict keys for use in the make_hab_risk_rasters function.


Former-commit-id: 38d2eea62827131072c8d192bed8a2c87e19578d