Update issue 1565
Created the function that will send to shapefile, and made a raster to base the shapefile on. This raster will be a 1 if there is high risk land there, or a 0 if there is nodata, or if tehre was data, but was MED or LOW risk.

Still need to call the raster utils function. Or just put into the core function.


Former-commit-id: afee71d14c979fe132d3f7b3bbb77850965e0b97