Changing global_dem URI to global_dem_meters
In a moment of confusion it was determined best to explicitly name the global
dem with its depth units. Thus it is now named in
invest-data/Base_Data/Marine/DEMs/ as global_dem_meters.

I think I have changed all paths in the invest-3 repository that needed to be
changed. Hopefully I didn't miss any.


Former-commit-id: 42d08d8fc67e398a2bf8127ba1681525fe869301