Udpate issue 1288
Created core test for 0 value ratings. However, I'm thinking that this should never actually happen. If we have an entry with a 0 value rating, we need to remove the entire dictionary, otherwise we are still incorporating the DQ and the W from that subdictionary into our equation without the rating to balance it out. Unless that's somehow what we want?


Former-commit-id: bf74ca180a9eefb7e037f7fd13fd779d3f9f8319