Update issue 1230
Was trying to index inproperly into the dictionaries. Instead, need to do a .values() or a .iteritems so that we get a viable dictionary to index into using the criteria name, or the 'Rating', 'DQ' or 'W". Value is fine when we don't need to know the name, but for the re-burning of the raster criteria, we needed to do an iteritems.


Former-commit-id: f57921b9db3295511bef865a41f02149f7b94eb9