Update issue 1585
The raster now being placed in the risk raster dictionary is now coming directly from one of the two risk calculation equations, rather than being an array returned and burned into a new raster band. Since we are passing in a base raster, should be able to return a correct raster according to the equation desired, and pass directly.


Former-commit-id: 866f1f01910ed33629cdb0ade6ca9ef96f36a247