Update issue 1585
Updated make_risk_rast_euc to use vectorize raster. However, need to actually burn it to a file within these particular functions. So, need to further modify to take in the uri. Can still return the ds to the overall function for storage for later, but want it to be done within the subfunctions instead.


Former-commit-id: 51ca68e64cb3361662834017552d1a9e1acf7116