Update issue 1597
Refactoring Biophysical Uri handler to retrieve hub_height from the CSV file
instead of as an input. In doing this, I had to change the structure of the code
around because I need the hub height to do the earliest operations. Thus I had
to move the CSV file handling to the top of the function.

I also added some comments and added another argument to be passed in the
dictionary. The argument is scale_key, which I end up computing again in the
core. Better practice to not duplicate and pass it in


Former-commit-id: 1bffe21a60c056d2a56f876fa11d903c056d3cb1