Fixes issue 1621
Changed code so that resampling is done first. It now takes about a second to perform compared to about 6 minutes. The problem came from the order in which the functions were called: I used to clip first, where the clipping function called vectorize_rasters, which resampled the raster by doing an expensive interpolation. I now resample first, which is much faster, and then clip.


Former-commit-id: 816490b36db0bbe5e74eec80c13d94dd8741c9a1