Update issue 1623
Thinking about habitat and mobile species as being two different things is ridiculous, and will cause problems. If we are going to have those as folders, we might as well have a stressor folder too, since having an 'Exposure' subfolder (stress-specific) withing a habitat makes little sense. Instead, have the second fodler layer just be 'Exposure', 'Resilience' or 'Sensitivity', so that the model can be looked at as less biased towards a habitat-focused POV.


Former-commit-id: 2a1f2636ae7fcf01e441c9cbd93e9d81c40fbaef