Update issue 1617
Fixed a bug where the projections were being tested only through equality of the WKTs.  This caused false negatives where shorthand names would be different even though the projections were identical.  Had to change to use the IsSame function in OSR and updated the test as well as the exception raised.


Former-commit-id: 4db2cbd72ac85f363813f6d19e645f93f3564787