Update issue 1626
The modifications to the model to generate Tiffs with vectorize_rasters caused some rasters to be of different sizes. This makes these resized rasters incompatible with the numpy functions that use them as arrays because of size issues. Fixed the code for this exact reason where the model computes coastal vulnerability no habitats. There are more places to fix, such as in compute_habitat_role.


Former-commit-id: fcb988789b9343c6be6dd82a03f348af7487e2da