Update issue 1598
Replacing my hacky attempt to check out a package version with the
official, pythonic method of using distribute to manage package
dependencies.

What this means is that when we discover a python package dependency, it
would behoove us to add that to the package requirements list in our
own setup.py.


Former-commit-id: 0c8f7339a28d1d60d4f14f711c413ce34e9829f3