Update issue 1583
Changed HRA non-core docstring to reflect that the intermediate outputs will be the non-DS included versions of the h-s/h/s dictionaries. The final outputs, which are the ones being passed on to core are the ones that will contain the 'DS' keys, and have reference to the burned individuals.


Former-commit-id: e333893c5889862a81005266b1fcdcac253e6024