Update issue 1267
Used Examples/rasterize_criteria_shapefiles to create the individual rasterized criteria shapefiles. Each raster will be burned with the value of attribute 'Rating' from the individual shapefiles.

This function should complete the set of burns to raster that we need to do as non-core. Once the program runs through the previously written functions (appear below make_crit_rast), we should have a completed dictionary that we can pass on to core. (But, need to re-write any dictionary use in core to do URIs, not open ds's.)


Former-commit-id: 1e6c222a764f0471648d2f25698227aa86fabf22