Update issue 1662
Turns out that fixing the segmentation fault is going to take much longer than expected. Because of time constraints, we'll keep this issue on the side, and fix other issues instead. I reverted the code to the pre-segmentation fault code, but substituted the function where the segfault happened (compute_segment_exposure) with the pre-vectorize_raster/vectorize_dataset version. I'll keep that issue priority as critical, though.


Former-commit-id: 4782deb131da5c96a5d1223aaec7503b7677f6fa