Update issue 1401
Finished the old pre proc write of the criteria into a csv. However, need to wait on adding any actual criteria names, since not sure how they're going to be coming in from the IUI.

Additionally, when they come in from the user, need to be doing our upper/lowercase check at that point.


Former-commit-id: 1a7a092998b4c6d7ddab076735fe4e5b9444e290