Update issue 1637
After talking to Greg, it seems much better to detect the habitats using the gdal rasterize technique referred to the 'ALL_TOUCHED=TRUE' option. It seem better to over-estimate the delineation of the habitats influence by less than a pixel than missing it altogether and having no zone of influence at all.


Former-commit-id: 4b7633c2aff83e6e92cd8cf1fddcd4612197f9cc