Update issue 1401
Added exception classes for the primary issues we want to be able to raise in the HRA preprocessor. These issues should not allow the model to proceed.

(Note: Should think over whether the < 4 and the improper spread are issues that we want to break model run on, or if we want to just somehow raise a warning, and allow the model to proceed.)


Former-commit-id: cc15eb4637d870263670af455b14f17e765fb9ab