Update issue 1668
The computation of the fetch is now bypassed and the model runs all through the end. Needed to reconstruct the fetch_distances and fetch_depths dictionaries from the data saved on disk. Also, fixed a bug in an assert that incorrectly used len() on a numpy array instead of '.size'


Former-commit-id: 0f5f311dabdb7bdeb6903fef18d91330fc62ff21