Update issue 1582
Fixing a bug in the DynamicPrimitive class where an element was only
validating when there was a validator and requirements were met.  Now,
validation is always called and recursive toggling is always toggled
when the value of a primitive is changed.

Also, this changeset fixes a bug in DynamicPrimitive's validate()
function where validate() was returning when self.root == None.


Former-commit-id: be4936c72cbbf0da4eba74468ca131e1d2329cdc