Update issue 1316
This issue was caused by the fact that validation was not happening when
an element's requirements were not met.  In the case of a text-field
based element (such as a file element), this means that validation does
not take place when there is no text.  This is not desired behavior, so
validation now happens whenever a change is made to the textfield.

Hopefully, this should also fix other issues in other model UIs.


Former-commit-id: 5fd43f0e503928f4f6bf87e30f371d9e8fd0eb4e