Update issue 1409
Fixed a bug where the fetch distances were saved in meters, but passed around in pixels. In fast mode, when the model reconstructed the fetch distance dictionary, the fetch distances were all in meters. Now, all the fetch distances are in meters throughout the model.

The model stores now the same boolean data about shore shallowness and shore enclosure that is used to determine shore exposure in the outputs.

Fetch distances and depths do not have a blue background anymore. Need to fix the function that adjusts the vulnerability ranking, though.


Former-commit-id: 8f82bc376cc9918098550653770e2ed3c82e2fab