Update issue 1574
Added JSON file output to the hra_preprocessor output folder. It will be stored in the same folder as the output CSV's, and be read by the non core when the user inputs the csv folder. It will be called "dir_names.txt" unless we actually end up storing something that isn't a dir name.


Former-commit-id: 9f6c294957ddb299059b5bdd5fb83221beaa3ef1