Update issue 1520
When checking for errors before running the model, errors are only
considered when the element is not disabled.  If the element is
disabled, its error messages are ignored.  This style of behavior
follows the precedent of the args_id, where the output value of an
element is fetched only if the element is not disabled.


Former-commit-id: a4699d231407cdb62ffb1df2dcd4a90895eece4c