Update issue 1722
Almost no raster have a background color set as nodata. This is in part due to the way set_save_info has handled no_data so far, where the function assigned a nodata value that was not in the raster. Instead, we want a large number of values to be nodata. This can be easily done by modifying the signature of set_save_info, which will be (once the API transition will be complete):
 - an optional no_data field can be specified by the user (set to None by default):
 - if no_data is set to None, compute a nodata value that is not already in the raster.
 - if no_data is not None, use this value

The function has a temporary awkward but non-API breaking signature to transition the code from the old to the new signature.


Former-commit-id: b9d5eb274877b920aca6c1fdf2664dd38d0a32c3