Update issue 1725
There are 3 parameters that come from the CSV files that are needed to calculate
the dimensions of the farm size for the farm size polyline output.

Theses parameters were being handled in the Valuation side, but now since we
moved it to the biophysical side, they needed to be handled there instead.


Former-commit-id: 1342122ef10b4df626a42efec97e3070487f0c45