Update issue 1761
I discovered that the issue was coming from passing around the datasource.
Changing everything to URI's will be amazing in the future.

When I was getting in the datasource to use to calculate distances, the pointer to
the features was to the end, thus distances were coming back zero because no
geometries were found. By calling ResetReading on the layer, it fixed the
problem


Former-commit-id: 779550b1452ae441bce5d35aa89cf4efb3cc3558