Update issue 1641
Fixed a bug where the depth threshold for function H in eq. 7 on the online notes was set 1000x too high so that it returned zeros all the time.
Fixed gravity constant from 0.981 to to 9.81.
Moved code that saves wave height and wave period to compute_wave_exposure.
Commented out lots of debug output.


Former-commit-id: d58161e836fe52c5c4d6c70dbd39de73e57ab4b8