Update issue 1763
The issue was occuring because the layer name was coming in as a unicode string. Actually, it was coming in as a numpy unicode string. We fixed it so that it's no longer a numpy unicode string, and was just a unicode string, but it was still having issues. We've now cast it to utf-8, so that it will use the smaller set of characters, HOWEVER, it is not a full unicode set. Definitely something to take a look at in the future.


Former-commit-id: 86c17028f6887736e2904e62ab0d58c22ce15439