Udpate issue 1245
Fixed an issue where the buffering around a stressor was being cut off early because of the size of the original shapefile envelope. Now I increase the envelope size for the new raster base, then buffer based on that size for the new buffered stressor raster.

Note: On my small example script, I was encountering an error where it was recognizing 0 as nodata, not zero by itself. Just make sure that when we do all overlap combinations, that we check for each layer's nodata, and set to zero if we find their nodata. Don't assume 0, even though I guess I call it specifically.


...maybe something to look at later.


Former-commit-id: c481bf1924258b0e6f7c880bce9f8a38a813fd47