Update issue 1787
Did some major overhaul in wave energy model.

I used vectorize_points instead of my own interpolation method. However doing
this provided an issue where the output was not clipped nicely to the AOI. Thus
I created a function that gets the convex hull of a point shapefile and creates
a polygon shapefile from that convex hull geometry. I then used this polygon to
clip the dataset from. I documented the new function but need to write tests for
it. Although in the end this function will be destroyed because when wave energy
gets refactored for putting bio and val together, it won't be needed.

I was able to remove my own interpolation method since nothing else was calling
it


Former-commit-id: b8a33aff12778d4763908c582a68a69bc87f08f2