NUMPY IS THE DEVIL.
(Need to make sure that when doing vectorize ops that the output of something is a float so that the whole raster is a float. Because apparently the GST_Float32 didn't make that clear enough. Though, this does bring up a concerning behavior that could occur if we never hit that if case. Should really test the same shapefile as both habitat and stressor and see what happens.)


Former-commit-id: 7252a7ee042a23c39fb8694bb13185bf4b4cc4ad