Update issue 1612
All the files that were formerly created in the ouputs directory are now created in the intermediate directory. The outputs directory will likely have individual single-band raster files, and a few multi-band rasters. Both will have exactly the same dimensions and the same data in them. This will be made easy with the function raster_utils.align_dataset_list that will clip, resize, align, and resample the data in the intermadiate directory and save it in outputs/. Then, it's going to be easy to create the fat multi-band rasters.


Former-commit-id: a8cff2d61ffbc3b0f24c5fb53fd12da365ad476a