Update issue 1612
Replaced spaces by underscores in the second element of each entry in output_info, so that key-related errors are easier to find in the code.

To make the creation of groups easier, the model is now saving all the new files created by compute_* and save_* in the args dictionary directly, and not in a separate args['uri'] entry. All the references to args['uri'] have been removed.

Due to the above design decision, all new entries created by the compute_* and save_* functions have to be unique and not erase already existing entries in args. Check for this and assert if new entries will erase existing ones in args.

Had to remove/rename some entries created in args to prevent name collision in the args dictionary. This takes the model closer to a uri only api.


Former-commit-id: 90fdc5e021ae6ea79e5fae3046b30c97b7d8145b