Update issue 1612
Changed the way the information to build the multi-band rasters is organized: I used to have a list of strings that I used to match keys in the args dictionary. I had too many files, because I would pick up any URI which key contains the string in the list. To solve the problem, I replaced the string by a tuple ('operator', 'token'). If the operator is 'include', any URI which key contains 'token' is selected. If the operator is 'match', only the URI which key exactly matches 'token' is valid.

Implemented the call to align_dataset_list and save all the output files in the outputs/ directory.
Created all the multi-band rasters.


Former-commit-id: 12cfa396606a6b4f0061c167dff71fff2159c5fa