Update issue 1830
The files created by the functions in execute and returned as URIs are copied to their respective directories and adjusted so that the files can be referred to from their new location.

Remove the files from the old location at the end, since they might be necessary and some dependencies are not moved, such as the 3 other files that are created besides the 'shp' shapefile.

Added some code to pick up the remaining files that are created in the pre-processing stage.


Former-commit-id: 4acfc3537f739c974b7e6e15e4ea4cdc7f9b261a