Update issue 1830
Once the remaining files that needed to be moved to their respective directories were pooled in a list, all elements of this list were checked all over for each group. Now using sets to remove the files that have already been moved over.

Use regex ibecause we want to match the prefix to the beginning of the filename, and not anywhere else.

Added some comments and output debug code.


Former-commit-id: 01aaaa7032a8730243aaca6929e1a3fb9b9ca5e6