Update issue 1564
Fixed issue with corruption of the AOi layer. It was being incorrectly validated within IUI as a "File" instead of "OGR". For some reason, this was causing it to wipe the .shp file. Properly checking now. No longer breaks everything.

(The other code changes are just debugs. But I'm lazy.)


Former-commit-id: 27bf7093485db855239d00ea91de37af3b92410d