Update issue 1564
Since we now know that aggregate takes in only numeric field identifiers, need to make a copy, and run the assessment with a new set of fields that are numbers. Also created a number map to the names so that it can be associated in the way the users intended.


Former-commit-id: bae8812892bc8b8ff60dd30a8421147cae381ef0