Update issue 1854
Refactoring build_utils a little to add the ability to change to a
different directory if requested.  This fixes an immediate issue of
needing to access the InVEST version information and write a file inside
tha package, but I believe that it will also help with resolving an
issue of how we should be able to use any mercurial project to get
invest-style version numbers.


Former-commit-id: 3e10bc8e696a61957ce99b02dc16a07be7c45f2e