Update issue 1851
The issue with RIOS IPA was due to an assumption I had made when
addressing issue 1316: that we always want to disable the container
itself when disabling the elements inside a container.  This is indeed
True, but only when the container is collapsible.

This changeset adds another conditional so that the container itself is
only also disabled when the container is collapsible.


Former-commit-id: f3f8e8ffce6aac422129a900fee18e72ec77c76c