Udpate issue 1564
Added max_risk back as a param to make_aoi tables, since we do actually need it for risk % calculation.

Additionally, looks like I forgot that I need an ecosystem raster too. Let's have that be returned as a separate data structure from pre_calc_avgs. It will take in the ecosystem ratser separately (have that raster return when it's calculated) and will do it after the for loop for H/S pairings. Will output to the same folder, single URI.


Former-commit-id: 3423ae2f457568c9debe058b60cbd2ecbdde1bbf