Update issue 1887
Intermediate commit.  This test will fail.

I'm trying to take the unprojected points shapefile I created from the
user's latlong coordinates and project it so that vectorize_points will
recognize that there are actually points to operate on.

Right now, vectorize_points does not recognize any points, since the
bounding boxes for the points are outside of the raster I have passed
in.


Former-commit-id: 5733f5ccd937f1f03dcabcbd56c51f7d0c8260fd