Update issue 1915
Restructured how the data is stored in the dictionary. Decided to make unique
keys for each point in a time step and then set that key to a dictionary which
is the entire rows information.

This allows me to get all the data I need, makes it expandable to future fields
and lets me use future functions down the road easier.

Adjusting test case to reflect change


Former-commit-id: a29638af118b67491290c68ff7cc51901957005d