Update issue 1230
Changing the structure of the call to aggregate_raster_values. Instead, we are now pre-creating all management zones within our dictionary, and steting their value as 0. This way, even if there is no overlap, we can still have a 0 value to show, instead of breaking the program.

I guess alternately, I could have just accepted this. But really, the dict.update() function is kind of neat.


Former-commit-id: 0bf3658f411e4bef5b08f646dcdb9709708ae9bf