Fixes issue 1958
Incorrect depths cause a warning message to be printed, and the values in the CSV are those actually used in the model fro wave height and wave period. Erroneous wind speeds (U <= 0) are set to 1 before they are used in the model, and a warning message is issued as well.


Former-commit-id: f05b0e9342f2ee632efe4817d6f99ee081e54700