Update issue 1888
Adding a test function with some scratch code for implementing the
flood water discharge function as a convolution.  I'm sure that I just
don't understand how to structure the problem to actually make this
happen as a convolution just yet ... and the problem is that a kernel is
fixed as we move across the landscape.  With the flood water discharge
function, the kernel would change with each new pixel, right??  When we
visit a new pixel, the state of the neighbor may change from not flowing
in to this pixel to actually flowing into this pixel.


Former-commit-id: 8fbcc197e5cf5b2b3934a59a0ee864b8eb2e4ea8