Update issue 1962
When using the continental shelf layer instead of the depth contour to determine the limits of the continental shelf, the contiental shelf layer from the shapefile doesn't get rasterized properly. Instead of the expected limits of the shelf, the raster is completely filled with 1s. When I use the exact same code on the global landmass polygon, the raster looks just fine. Swapping the code that rasterizes the landmass polygon with the code that rassterizes the continental shelf.


Former-commit-id: f07e1b99b4bedc1c040e8cf9679b6fb7b549db5f