Update issue 1784
Restructuring the way that we check file permissions in iui_validator.
If the file or folder does not exist, iui_validator will automatically
check the parent folder instead.  It is up to the user to define in the
JSON if the folder is required to exist before permissions are checked.


Former-commit-id: 6cc5031ae54ac41de83f4db7f86c763858b5c873