Update issue 1962
The model completes, without segfaults. Made many changes all over so the model completes successfully. The output doesn't look right, though because of the contiental shelf. The polygon in the global shapefile does not go allthe way to the shore. Instead, according to Greg Guannel, the shelf ends at the polygon's edge that is closest to the shore. So I should detect the polygon edges and use the colsest one to assess the distance from the shore to the end of the shelf.


Former-commit-id: c38505133726ce0d87153c7b31c3453fa34c209c