Update issue 1997
Added debug code to find out that the model chokes because the bathymetry and the AOI are not align and only partially overlap. Will have to use align_rasters with memory_mapped arrays, and probably refactor the pre-processing function to get nicely aligned and overlapping rasters before the computation begins.


Former-commit-id: 5f52806c9406f2357b13bf50ef99a64984bf4e91