Update issue 2020
Making some major changes. Created some new functions that add a new field/val
to a shapefile based on a dictionary. This is to add fractp_mn and other values
instead of creating unneccessary raster outputs for them.

Also added a function that computes the wyield volume and volume for ha. This
eliminates some output rasters that are just polygons anyway


Former-commit-id: d505b678aca40f23fc909746bb0377fadbaa2899