Update issue 1973
Created a new function that adds a row to a CSV file, or creates a new CSV file
and writes that first row.

Creating different CSV files for each water / sub-watershed. Storing the URIs in
a dictionary that map to the keys of the sheds.

Looping over the data monthly to build up the proper dictionar to then add to
the end of the CSV files appropriately


Former-commit-id: d1a91353443aed864048fed45e584f5c8aac7254