Update issue 1956
Changing how the alpha values are being input. I did just have them hard coded,
but now they come in as a CSV file. Created a function to unpack the CSV and
load it into a dictionary like I had set it before.

The Beta value is now also input through this table


Former-commit-id: 7a452d7b01eeb3cf3f2610c1b48b338a4a6f9d6b