Update issue 2150
Now calculate the water yield volume by using number of pixels and pixel area.

Also now calculate the hectare volume using the hectare mean.

This required adding an argument to the wyield volume function for pixel area.
Also had to break out aggregating over the water yield raster from it's
function. This is because I wanted the number of pixels for only this raster and
added that value to the sub/watershed shapefiles


Former-commit-id: 8d32d9b1f9fdb8fb284888920bce8b98ed153c81