Update issue 2163
A bunch of things had gotten changed since the official "release" of the HRA model. We had tried to fix it to shoehorn it into new functionality, and ended up breaking some things without realizing it. Fixed issues with regex searches, accidentally deleted directories, overdeleting directories, and a file that was just not saving as',Memory'. We're changing that one to just be ESRI Shapefile, and delete after the fact.

All 3 parts of HRA (two UI's) should be a working copy now, but I stress: IT SHOULD NOT BE USED FOR ACTUAL CALULATIONS. We will hopefully be releasing the corrected HRA soon, but in the meantime, this can be used for demonstrations on the HRA_3_0 sample data. I'll post a link into this issue of a link to the nightly build version that you can download for demos at trainings.


Former-commit-id: 8da612db9a576ca32fd9208ec1f0485d906d66d9