Update ixxue 2179
THe problem occured when the absorption rate was undefined. This rarely, but legitimately happens on edges where a slope can't be calculated.  Strange, but we can break the routing if we ignore it completely.

I've set the absportion rate to 0.0 if we encoutner a nodata value in a region that otherwise has well defined data.


Former-commit-id: de54ffc0577ad39dbf9e13458581dfec917badd4