Update issue 2157
There was a bug in the function that reads in the precipitation and ETo time
step CSV files. During one loop I was iteraterating over the length of a String
and not the length of the list... Ooops.

Somehow this didn't error before because my list and string length were the
same. Crazy.


Former-commit-id: 78051063c6acc95882257f66caa331f873edb99a