Update issue 2063
What we really want to do when we add a pixel stack is to treat any nodata as if they're 0 when we do the mass pixel adds. This shoudl search first for nodata, if all nodata, then return nodata, otherwise, add the pixel value as long as it isn't nodata.


Former-commit-id: 4eba5e30b6f379857e4aae7df21c71191dfa5746