Update issue 2044
This is an awfully big commit and should be code reviewed by Adam... Joking.

This dump is a hypothetical completion of the wind energy biophysical refactor.
All the code from the core and biophysical sides have been collected and
integrated into proper URI format.

I also added a few functions from core as well as created some new functions /
functionality.

The biggest to note is that I wrote a function which returns the geometry for
for the wind energy point that has the highest harvested value. This is to use
in creating the box for the output size of the farm. Hopefully this will better
position the farm size output.

In my attempt to make everything URI based I added an inner function for looping
over the wind energy points and computing the harvested energy and density.


Former-commit-id: c1339d02004c86e736e1f454bc1f95fe92f0b6c8