Update issue 1731
Rob has requested that we use bilinear interpolation for the wind energy model.
Therefore this issue is relevant and is where I will dump the commits for.

Note, that in issue 2044 I am doing a URI refactor of Wind Energy and have
already made some changes for interpolation over there. I was planning on
rolling it out altogether, but Rob would like the interpolation stuff a little
sooner than I think I could get out the door with the whole refactor.

Therefore in this commit I am adding an interpolation parameter to the
vectorize_points call, passing in 'linear'. We use 'linear' because it will act
like bilinear on a 2D array. In issue 2044 I made the necessary changes to
vectorize_points in raster_utils to take this new optional argument.

When testing this new interpolation scheme it seems that Scipy 0.12.0 is needed.
When running it using Scipy 0.11.0 I get an error in scipy.interpolate.griddata
of ZeroDivisionError. I updated to 0.12.0 and it works fine and the results seem
reasonable.


Former-commit-id: c192565ea47eaf988ade028e142ee627694d7190